requirements.txt
flask
pyyaml
markdown-it-py
linkify-it-py